+++ addressbook/src/main/java/org/jugs/webdav/addressbook/AddressBookApplication.java
@@ -27,7 +27,7 @@ import java.util.logging.Logger;
 import javax.ws.rs.core.Application;
 import javax.xml.bind.JAXBException;
 
-import net.java.dev.webdav.jaxrs.xml.WebDavContextResolver;
+import org.jugs.webdav.jaxrs.xml.WebDavContextResolver;
 
 /**
  * Sole JAX-RS Application of JPA Address Book Sample.