+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+* dependency to old `net.java.dev.webdav` dependency completely removed
+
+
 ## [2.3] - 2022-08-05
 
 * [addressbook](addressbook/README.html) addes as new module